requirements.txt
python-telegram-bot==21.5
httpx==0.27.2
requests==2.31.0
Flask==3.0.3
